Expand readme and add a contributing file.
Added more detail to the readme including a summary. Added detail to the
git flow to reflect the previous comments I had made about using
branches. Added a contributing file which will need more love but is
often where folks start when they look at contributing to a new project.
Moved the todos for the readme down to the bottom for later.